requirements.txt
fastapi>=0.115.0
openai>=1.68.0
pydantic>=2.10.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
uvicorn[standard]>=0.34.0
numpy>=1.26.0
pytest>=8.3.0
